Fixed the `extract_boxed` method in `local_tester.py`.
We found that some models might generate an empty `\boxed{}`
without an answer in the middle of a solution, even if they
finally solve the problem correctly in a later `\boxed{}`.
Therefore, we have updated the method to extract the content
from the *last* occurrence of `\boxed{...}` to improve stability.

Also, we suggest using more robust tools, such as harness,
for evaluation, as this script is relatively simple.